# openApi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/mock/user/getUserList": {
      "get": {
        "operationId": "UserController_getUserList",
        "parameters": [
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "一页多少条",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "第几页",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "nickName",
            "required": false,
            "in": "query",
            "description": "昵称",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "username",
            "required": false,
            "in": "query",
            "description": "账号",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/UserInfoVO"
                              }
                            },
                            "total": {
                              "type": "number",
                              "default": 0
                            },
                            "pageSize": {
                              "type": "number",
                              "default": 0
                            },
                            "PageNum": {
                              "type": "number",
                              "default": 0
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/mock/user/createUserInfo": {
      "post": {
        "operationId": "UserController_createUserInfo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/mock/user/login": {
      "post": {
        "operationId": "UserController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/mock/user/findUserInfo": {
      "get": {
        "operationId": "UserController_findUserInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/mock/user/getCaptcha": {
      "get": {
        "operationId": "UserController_getCaptcha",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取验证码",
            "content": {
              "image/svg+xml": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/mock/health/network": {
      "get": {
        "operationId": "HealthController_checkNetwork",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health - 健康检查"
        ]
      }
    },
    "/mock/health/memory": {
      "get": {
        "operationId": "HealthController_checkMemory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health - 健康检查"
        ]
      }
    },
    "/mock/health/db": {
      "get": {
        "operationId": "HealthController_checkDB",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health - 健康检查"
        ]
      }
    },
    "/mock/health/disk": {
      "get": {
        "operationId": "HealthController_checkDisk",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Health - 健康检查"
        ]
      }
    }
  },
  "info": {
    "title": "Mock API",
    "description": "The Mock API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "MOCK",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ResponseModel": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "code": {
            "type": "number",
            "default": 200
          },
          "message": {
            "type": "string",
            "default": "success"
          }
        },
        "required": [
          "data",
          "code",
          "message"
        ]
      },
      "UserInfoVO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "账号"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "role": {
            "type": "string",
            "nullable": true,
            "description": "角色 USER|ADMIN",
            "default": "USER"
          },
          "nick_name": {
            "type": "string",
            "description": "昵称",
            "default": "昵称"
          },
          "avator": {
            "type": "string",
            "default": "头像"
          },
          "sex": {
            "type": "string",
            "default": "UNKNOW"
          }
        },
        "required": [
          "username",
          "password",
          "role",
          "nick_name",
          "avator",
          "sex"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "账号",
            "default": "账号"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "default": "密码"
          },
          "role": {
            "type": "string",
            "nullable": true,
            "example": "角色 USER|ADMIN",
            "default": "USER"
          },
          "nick_name": {
            "type": "string",
            "example": "昵称"
          },
          "avator": {
            "type": "string",
            "example": "头像"
          },
          "sex": {
            "type": "string",
            "example": "UNKNOW"
          }
        },
        "required": [
          "username",
          "password",
          "role",
          "nick_name",
          "avator",
          "sex"
        ]
      },
      "LoginUserDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "账号",
            "default": "账号"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "default": "密码"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "账号"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "role": {
            "type": "string",
            "description": "角色权限"
          },
          "nick_name": {
            "type": "string",
            "description": "昵称"
          },
          "avator": {
            "type": "string",
            "description": "头像"
          },
          "sex": {
            "type": "string",
            "description": "性别"
          },
          "create_at": {
            "type": "string",
            "description": "创建时间"
          },
          "update_at": {
            "type": "string",
            "description": "更新时间"
          },
          "delete_at": {
            "type": "string",
            "description": "删除时间"
          }
        },
        "required": [
          "username",
          "password",
          "role",
          "nick_name",
          "avator",
          "sex",
          "create_at",
          "update_at",
          "delete_at"
        ]
      }
    }
  }
}